Fixed more spelling issues and clarified rotation initializer
