--- LeonidsLib/src/main/java/com/plattysoft/leonids/initializers/RotationInitiazer.java
@@ -1,24 +1,0 @@
-package com.plattysoft.leonids.initializers;
-
-import java.util.Random;
-
-import com.plattysoft.leonids.Particle;
-
-public class RotationInitiazer implements ParticleInitializer {
-
-	private int mMinAngle;
-	private int mMaxAngle;
-
-	public RotationInitiazer(int minAngle, int maxAngle) {
-		mMinAngle = minAngle;
-		mMaxAngle = maxAngle;
-	}
-
-	@Override
-	public void initParticle(Particle p, Random r) {
-		int value = mMinAngle == mMaxAngle ? mMinAngle : r.nextInt(mMaxAngle-mMinAngle)
-				+mMinAngle;
-		p.mInitialRotation = value;
-	}
-
-}
